Bugfixing exporter mockserver testcases (#10)
This commit fixes current bugs within the VCSA-Exporter, the testcases and the mock server.
VCSA-Exporter: fixed inconsistencies in return values of the connection class and a regression bug which caused the metric attribute to grow in size the more requests are made against it.
testcases: fixed the consecutive runs test. The previous version caused issues in comparing the output length of the VCSA-Exporter. The Prometheus header entries were now removed in order to compare the metrics data appropriately.
Also fixed the test_logout_failure method. It tested against a wrong password instead of correctly testing against the sessionID.
MockServer: fixed the delete method in the Authentication Class. The previous version always returned a 200 status code instead of first checking against the sessionID.